submit

+++ server/src/main/java/atmin/common/event/PaymentSuccessEvent.java
@@ -1,11 +1,12 @@
 package atmin.common.event;
 
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 public record PaymentSuccessEvent(
         Long orderCode,
         String transactionReference,
         String invoiceNumber,
         LocalDateTime issuedDate,
-        Double totalAmount
+        BigDecimal totalAmount
 ) {}
