+++ server/src/main/java/atmin/modules/product/dto/ProductDto.java
@@ -5,6 +5,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+import java.math.BigDecimal;
 import java.util.Map;
 
 @Data
@@ -15,7 +17,7 @@ public class ProductDto {
     private String id;
     private String name;
     private String category;
-    private Double price;
+    private BigDecimal price;
     private String material;
     private Double rating;
     private Integer reviews;
